feat(ui): remove unused components
