--- frontend/app/components/Session_/Fetch/index.js
@@ -1,1 +1,0 @@
-export { default } from './Fetch';
