--- frontend/app/components/Session_/Network/network.module.css
@@ -1,31 +1,0 @@
-
-
-.location {
-  min-height: 32px;
-  display: flex;
-  align-items: center;
-  padding: 5px 10px;
-  font-size: 12px;
-  line-height: 12px;
-
-  & > div:last-child {
-    font-weight: 300;
-    max-width: 80%;
-    overflow: hidden;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-    padding: 5px 0;
-  }
-}
-
-.popupNameTrigger {
-  white-space: nowrap;
-  text-overflow: ellipsis;
-  overflow: hidden;
-  max-width: 80%;
-  width: fit-content;
-}
-.popupNameContent {
-  max-width: 600px;
-  overflow-wrap: break-word;
-}
